server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.3
    # via redis
attrs==23.1.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==23.6.2
    # via daphne
automat==22.10.0
    # via twisted
cffi==1.15.1
    # via cryptography
channels[daphne]==4.0.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.1.0
    # via -r requirements.in
constantly==15.1.0
    # via twisted
cryptography==41.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.0.0
    # via channels
django==4.2.4
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.2.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.4
    # via
    #   hyperlink
    #   twisted
incremental==22.10.0
    # via twisted
iniconfig==2.0.0
    # via pytest
msgpack==1.0.5
    # via channels-redis
packaging==23.1
    # via pytest
pluggy==1.3.0
    # via pytest
psycopg2==2.9.7
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pyopenssl==23.2.0
    # via twisted
pytest==7.4.0
    # via
    #   pytest-asyncio
    #   pytest-django
pytest-asyncio==0.21.1
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
redis==5.0.0
    # via channels-redis
service-identity==23.1.0
    # via twisted
six==1.16.0
    # via automat
sqlparse==0.4.4
    # via django
twisted[tls]==23.8.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.7.1
    # via twisted
zope-interface==6.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
